+++ tests/testthat/test-bas.lm.R
@@ -105,3 +105,31 @@ test_that("GLM logit probne0", {
   expect_equal(pima.BAS$probne0, pima.BAS$probne0)
 
 })
+
+test_that("force.heredity", {
+  loc = system.file("testdata", package="BAS")
+  d = read.csv(paste(loc, "JASP-testdata.csv", sep="/"))
+
+  simpleFormula = as.formula("contNormal ~ contGamma + contcor1 + contGamma * contcor1 ")
+
+  set.seed(1)
+  basObj = bas.lm(simpleFormula,
+                  data = d,
+                  alpha = 0.125316,
+                  prior = "JZS",
+                  include.always=as.formula("contNormal ~ contcor1"),
+                  modelprior=beta.binomial(1,1),
+                  weights = d$facFifty)
+  set.seed(1)
+  basObj.old = bas.lm(simpleFormula,
+                      data = d,
+                      alpha = 0.125316,
+                      prior = "JZS",
+                      include.always=as.formula("contNormal ~ contcor1"),
+                      modelprior=beta.binomial(),
+                      weights = d$facFifty, force.heredity = FALSE)
+  basObj.old = force.heredity.bas(basObj.old)
+
+  expect_equal(basObj$probne0, basObj.old$probne0)
+
+})